LINEBOTPJ-124 重構專案結構 1. 移除strategy.py、template.py、task.py，改新增features資料夾存放每個功能的py，並在該py用@register_feature裝飾器註冊該功能，並實作execute_message()及execute_postback()方法 2. config.py 使用動態檢查環境變數
