+++ config.py
@@ -23,44 +23,39 @@ class Singleton(type):
     
 class Config(metaclass=Singleton):
     def __init__(self):
-        self.CHANNEL_SECRET = os.getenv('CHANNEL_SECRET', None)
-        self.CHANNEL_ACCESS_TOKEN = os.getenv('CHANNEL_ACCESS_TOKEN', None)
-        self.SPREADSHEET_URL = os.getenv('SPREADSHEET_URL', None)
-        self.FIREBASE_CREDENTIALS = os.getenv('FIREBASE_CREDENTIALS', None)
-        self.LINE_NOTIFY_CLIENT_ID = os.getenv('LINE_NOTIFY_CLIENT_ID', None)
-        self.LINE_NOTIFY_CLIENT_SECRET = os.getenv('LINE_NOTIFY_CLIENT_SECRET', None)
-        self.LINE_NOTIFY_GROUP_TOKEN = os.getenv('LINE_NOTIFY_GROUP_TOKEN', None)
-        self.check_env()
-        self.line_bot_init()
-        self.feature_init()
+        self._load_environment_variables()
+        self._check_env()
+        self._line_bot_init()
+        self._feature_init()
 
-    def check_env(self):
-        """確認環境變數是否正確設定"""
-        if self.CHANNEL_SECRET is None or self.CHANNEL_ACCESS_TOKEN is None:
-            print("Please set LINE_CHANNEL_SECRET, LINE_CHANNEL_ACCESS_TOKEN environment variables.")
-            sys.exit(1)
+    def _load_environment_variables(self):
+        self.CHANNEL_SECRET = os.getenv('CHANNEL_SECRET')
+        self.CHANNEL_ACCESS_TOKEN = os.getenv('CHANNEL_ACCESS_TOKEN')
+        self.SPREADSHEET_URL = os.getenv('SPREADSHEET_URL')
+        self.FIREBASE_CREDENTIALS = os.getenv('FIREBASE_CREDENTIALS')
+        self.LINE_NOTIFY_CLIENT_ID = os.getenv('LINE_NOTIFY_CLIENT_ID')
+        self.LINE_NOTIFY_CLIENT_SECRET = os.getenv('LINE_NOTIFY_CLIENT_SECRET')
+        self.LINE_NOTIFY_GROUP_TOKEN = os.getenv('LINE_NOTIFY_GROUP_TOKEN')
+        self.GDRIVE_API_CREDENTIALS = os.getenv('GDRIVE_API_CREDENTIALS')
+        self.LIFF_ID_COMPACT = os.getenv('LIFF_ID_COMPACT')
+        self.LIFF_ID_TALL = os.getenv('LIFF_ID_TALL')
+        self.LIFF_ID_FULL = os.getenv('LIFF_ID_FULL')
 
-        if os.getenv('GDRIVE_API_CREDENTIALS') is None:
-            print("Please set GDRIVE_API_CREDENTIALS environment variable.")
-            sys.exit(1)
+    def _check_env(self):
+        required_vars = [
+            'CHANNEL_SECRET', 'CHANNEL_ACCESS_TOKEN', 'GDRIVE_API_CREDENTIALS',
+            'SPREADSHEET_URL', 'FIREBASE_CREDENTIALS', 'LINE_NOTIFY_CLIENT_ID',
+            'LINE_NOTIFY_CLIENT_SECRET', 'LINE_NOTIFY_GROUP_TOKEN',
+            'LIFF_ID_COMPACT', 'LIFF_ID_TALL', 'LIFF_ID_FULL'
+        ]
         
-        if self.SPREADSHEET_URL is None:
-            print("Please set SPREADSHEET_URL environment variable.")
-            sys.exit(1)
+        missing_vars = [var for var in required_vars if getattr(self, var) is None]
         
-        if self.FIREBASE_CREDENTIALS is None:
-            print("Please set FIREBASE_CREDENTIALS environment variable.")
+        if missing_vars:
+            print(f"Please set the following environment variables: {', '.join(missing_vars)}")
             sys.exit(1)
 
-        if self.LINE_NOTIFY_CLIENT_ID is None or self.LINE_NOTIFY_CLIENT_SECRET is None or self.LINE_NOTIFY_GROUP_TOKEN is None:
-            print("Please set LINE_NOTIFY_CLIENT_ID, LINE_NOTIFY_CLIENT_SECRET or LINE_NOTIFY_GROUP_TOKEN environment variables.")
-            sys.exit(1)
-
-        if os.getenv('LIFF_ID_COMPACT') is None or os.getenv('LIFF_ID_TALL') is None or os.getenv('LIFF_ID_FULL') is None:
-            print("Please set LIFF_ID_COMPACT, LIFF_ID_TALL, LIFF_ID_FULL environment variables.")
-            sys.exit(1)        
-
-    def line_bot_init(self):
+    def _line_bot_init(self):
         """初始化LINE Bot相關物件"""
         self.handler = WebhookHandler(self.CHANNEL_SECRET)
         self.configuration = Configuration(access_token=self.CHANNEL_ACCESS_TOKEN)
@@ -68,7 +63,7 @@ class Config(metaclass=Singleton):
         self.firebaseService = FireBaseService(json.loads(self.FIREBASE_CREDENTIALS))
         self.lineNotifyService = LineNotifyService(self.LINE_NOTIFY_CLIENT_ID, self.LINE_NOTIFY_CLIENT_SECRET)
     
-    def feature_init(self):
+    def _feature_init(self):
         self.feature = {
             'menu': FeatureStatus.ENABLE,
             'setting': FeatureStatus.ENABLE,
