+++ features/__init__.py
@@ -1,0 +1,12 @@
+from .base import FeatureFactory
+from .menu import Menu
+from .setting import Setting
+from .course import Course
+from .certificate import Certificate
+from .counseling import Counseling
+from .community import Community
+from .equipment import Equipment
+from .quiz import Quiz
+from .faq import FAQ
+
+__all__ = ['FeatureFactory', 'Menu', 'Setting', 'Course', 'Certificate', 'Counseling', 'Community', 'Equipment', 'Quiz', 'FAQ']
